+++ src/backend/accounts/settings.py
@@ -85,6 +85,7 @@ class Base(AuthenticationSettings, OIDCProviderSettings, Configuration):
     SECRET_KEY = SecretFileValue(None)
     SECRET_KEY_FALLBACKS = SecretFileListValue([])
     SALT_KEY = SecretFileListValue(None)  # django-fernet-encrypted-fields
+    SESSION_STATE_KEY_LENGTH = values.PositiveIntegerValue(32)
 
     # Application definition
     ROOT_URLCONF = "accounts.urls"
@@ -361,6 +362,9 @@ class Base(AuthenticationSettings, OIDCProviderSettings, Configuration):
     )
     FRONTEND_JS_URL = values.Value(
         None, environ_name="FRONTEND_JS_URL", environ_prefix=None
+    )
+    FRONTEND_LOGOUT_URL = values.Value(
+        None, environ_name="FRONTEND_LOGOUT_URL", environ_prefix=None
     )
     THEME_CUSTOMIZATION_FILE_PATH = values.Value(
         os.path.join(BASE_DIR, "accounts/configuration/theme/default.json"),