+++ src/frontend/apps/accounts/src/features/auth/conf.ts
@@ -2,4 +2,3 @@ import { baseApiUrl } from "@/features/api/utils";
 
 export const LOGIN_URL = `${baseApiUrl()}login/`;
 export const LOGOUT_URL = `${baseApiUrl()}logout/`;
-export const OIDC_LOGOUT_URL = `${baseApiUrl()}o/logout/`;