+++ CHANGELOG.md
@@ -21,5 +21,6 @@ and this project adheres to
 - 👽(authentication) handle Social Auth's login views requiring POST requests
 - 🦖(OIDC Provider) make introspection endpoint fall back to PSA backends
 - ✨(OIDC Provider) relay RP-Initiated logout confirmation to the frontend
+- ✨(auth) also logout from the upstream OIDC Provider when logging out
 
 [unreleased]: https://github.com/suitenumerique/accounts/compare/main